Seperate components into folders for maintainability
