+++ frontend/src/components/select/SelectCorridorCategory.tsx
@@ -6,11 +6,11 @@ import DialogContent from '@material-ui/core/DialogContent';
 import DialogTitle from '@material-ui/core/DialogTitle';
 import differenceWith from 'lodash/differenceWith';
 import { useEffect, useState } from 'react';
-import DB from '../DB';
-import { CorridorCategory } from '../models/CorridorCategory';
-import { compareByID } from '../utils/util';
-import NameList from './common/NameList';
-import CorridorCategoryEditor from './CorridorCategoryEditor';
+import DB from '../../DB';
+import { CorridorCategory } from '../../models/CorridorCategory';
+import { compareByID } from '../../utils/util';
+import NameList from '../common/NameList';
+import CorridorCategoryEditor from '../CorridorCategoryEditor';
 
 type Props = {
   open: boolean;
