+++ frontend/src/components/select/SelectItem.tsx
@@ -6,42 +6,42 @@ import DialogContent from '@material-ui/core/DialogContent';
 import DialogTitle from '@material-ui/core/DialogTitle';
 import differenceWith from 'lodash/differenceWith';
 import { useEffect, useState } from 'react';
-import DB from '../DB';
-import { Monster } from '../models/Monster';
-import { compareByID } from '../utils/util';
-import NameList from './common/NameList';
-import MonsterEditor from './MonsterEditor';
+import DB from '../../DB';
+import { Item } from '../../models/Item';
+import { compareByID } from '../../utils/util';
+import NameList from '../common/NameList';
+import ItemEditor from '../ItemEditor';
 
 type Props = {
   open: boolean;
-  exclude: Monster[];
-  onSelect: (m: Monster) => void;
+  exclude: Item[];
+  onSelect: (item: Item) => void;
   onCancelClick: () => void;
 }
 
-export default function SelectMonster(props: Props) {
-  const [monsters, setMonsters] = useState<Monster[]>([]);
-  const [monsterEditorOpen, setMonsterEditorOpen] = useState(false);
+export default function SelectItem(props: Props) {
+  const [Items, setItems] = useState<Item[]>([]);
+  const [ItemEditorOpen, setItemEditorOpen] = useState(false);
   const [isLoading, setIsLoading] = useState(false);
 
   useEffect(() => {
     let mounted = true;
     setIsLoading(true);
-    DB.getAllMonsters().then(result => {
+    DB.getAllItems().then(result => {
       if (mounted) {
         setIsLoading(false);
         if (result && result.valid) {
-          var list = differenceWith(result.response, props.exclude, compareByID) as Monster[]
-          setMonsters(list)
+          var list = differenceWith(result.response, props.exclude, compareByID) as Item[]
+          setItems(list)
         }
       }
     })
     return () => { mounted = false };
   }, []);
 
-  const handleSave = (cc: Monster) => {
-    setMonsterEditorOpen(false);
-    props.onSelect(cc);
+  const handleSave = (item: Item) => {
+    setItemEditorOpen(false);
+    props.onSelect(item);
   }
 
   return (
@@ -50,15 +50,15 @@ export default function SelectMonster(props: Props) {
         open={props.open}
         aria-labelledby="form-dialog-title"
       >
-        <DialogTitle id="form-dialog-title">Select Monster</DialogTitle>
+        <DialogTitle id="form-dialog-title">Select Item</DialogTitle>
         <DialogContent>
           {isLoading && 
             <div style={{textAlign: "center"}}>
               <CircularProgress/>
             </div>
           }
-          <NameList<Monster> list={monsters} onClick={(m: Monster) => props.onSelect(m)}></NameList>
-          <Button onClick={() => setMonsterEditorOpen(true)} variant="outlined" style={{ width: "100%" }} color="primary">
+          <NameList<Item> list={Items} onClick={(item: Item) => props.onSelect(item)}></NameList>
+          <Button onClick={() => setItemEditorOpen(true)} variant="outlined" style={{ width: "100%" }} color="primary">
             Add New
           </Button>
         </DialogContent>
@@ -68,11 +68,11 @@ export default function SelectMonster(props: Props) {
           </Button>
         </DialogActions>
       </Dialog>
-      {monsterEditorOpen &&
-        <MonsterEditor
-          open={monsterEditorOpen}
+      {ItemEditorOpen &&
+        <ItemEditor
+          open={ItemEditorOpen}
           onSave={handleSave}
-          onCancelClick={() => setMonsterEditorOpen(false)}
+          onCancelClick={() => setItemEditorOpen(false)}
         />
       }
     </div>
